python/pyproject.toml
[tool.poetry]
name = "kalendar"
version = "0.7.0"
description = ""
authors = ["netusil <m.netusil@mikrovox.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.7.*"
rpi-backlight = "^2.6.0"
kivy = {extras = ["base"], version = "^2.1.0"}
python-dotenv = "^0.21.0"
pydantic = "^2.5.1"
"kivy-garden.graph" = {url = "https://github.com/kivy-garden/graph/archive/master.zip"}
pytz = "^2024.1"


[tool.poetry.dev-dependencies]
mypy = "^1.4.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
